+++ src/main/java/org/observertc/webrtc/observer/evaluators/PCSObserver.java
@@ -8,7 +8,7 @@ import org.observertc.webrtc.observer.ObserverConfig;
 import org.observertc.webrtc.observer.common.IPAddressConverterProvider;
 import org.observertc.webrtc.observer.dto.AbstractPeerConnectionSampleVisitor;
 import org.observertc.webrtc.observer.dto.PeerConnectionSampleVisitor;
-import org.observertc.webrtc.observer.dto.v20200114.PeerConnectionSample;
+import org.observertc.webrtc.observer.dto.pcsamples.v20200114.PeerConnectionSample;
 import org.observertc.webrtc.observer.monitors.ObserverMetrics;
 import org.observertc.webrtc.observer.samples.ObservedPCS;
 import org.slf4j.Logger;